+++ __audits__/lighthouse/lhreport__20210803.csv
@@ -1,0 +1,7 @@
+, URL, Mobile_Performance, Mobile_Accessibility, Mobile_Best_Practices, Mobile_SEO, Desktop_Performance, Desktop_Accessibility, Desktop_Best_Practices, Desktop_SEO,
+,https://revenuedata.doi.gov/,66,90,100,97,75,90,100,97,
+,https://revenuedata.doi.gov/explore/,49,86,79,95,53,85,79,95,
+,https://revenuedata.doi.gov/query-data/,59,92,86,100,54,92,79,100,
+,https://revenuedata.doi.gov/downloads/,86,100,86,100,80,100,86,100,
+,https://revenuedata.doi.gov/glossary/,80,84,86,91,73,84,86,91,
+,https://revenuedata.doi.gov/how-revenue-works/,86,99,86,100,85,99,86,100
