--- __tests__/lighthouse/lhreport__20210803.csv
@@ -1,7 +1,0 @@
-, URL, Mobile_Performance, Mobile_Accessibility, Mobile_Best_Practices, Mobile_SEO, Desktop_Performance, Desktop_Accessibility, Desktop_Best_Practices, Desktop_SEO,
-,https://revenuedata.doi.gov/,52,90,100,97,65,90,100,97,
-,https://revenuedata.doi.gov/explore/,67,85,79,95,50,86,79,95,
-,https://revenuedata.doi.gov/query-data/,53,92,79,100,49,92,79,100,
-,https://revenuedata.doi.gov/downloads/,81,100,86,100,66,100,86,100,
-,https://revenuedata.doi.gov/glossary/,69,84,86,91,51,84,86,91,
-,https://revenuedata.doi.gov/how-revenue-works/,62,99,86,100,82,99,86,100
